Add no-data-to-display as an always included module (cdnScriptsCommon)
So users are able to export empty charts with the default message. Useful when
converting all charts to a static image for example for a download, dump or
(PDF) report. Without this module you would get empty images (making it unclear
something went wrong or this is intended), with this module you get a clear "No
data to display" message.

Fixes:
- https://github.com/highcharts/node-export-server/issues/151
- https://github.com/highcharts/node-export-server/issues/257